+++ sdk/src/main/java/com/mapbox/maps/viewannotation/ViewAnnotation.kt
@@ -23,6 +23,11 @@ internal data class ViewAnnotation(
    */
   var visible: Boolean,
   /**
+   * Cached value if view is actually shown. View could already be visible (and be treated as visible by core)
+   * but it takes some time to actually show it on top of the map.
+   */
+  var shown: Boolean?,
+  /**
    * Layout params needed to position Android view correctly on the screen.
    */
   var viewLayoutParams: FrameLayout.LayoutParams,